requirements.txt
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.2
anyio==4.4.0
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
argcomplete==1.8.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.1.0
blinker==1.9.0
bloom==0.12.0
breezy==3.2.1
Brlapi==0.8.3
Brotli==1.0.9
CacheControl==0.14.2
cachetools==5.5.1
captcha==0.7.1
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2020.6.20
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.1.7
cloudpickle==3.0.0
cmdtest==0.32+git
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
contourpy==1.1.0
cov-core==1.15.0
coverage==6.2
crudini==0.9.3
cryptography==45.0.5
cupshelpers==1.0
cycler==0.11.0
Cython==3.0.10
dask==2024.8.0
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.8
distlib==0.3.8
distro==1.7.0
docutils==0.17.1
dulwich==0.20.31
empy==3.3.4
exceptiongroup==1.2.2
executing==2.0.1
extra-streamlit-components==0.1.80
eyeD3==0.8.10
fastbencode==0.0.5
fastimport==0.9.14
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
firebase-admin==6.6.0
flake8==4.0.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.42.0
fqdn==1.5.1
fs==2.4.12
fsspec==2024.6.1
future==1.0.0
gast==0.5.2
geant4-pybind==0.1.0b13
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
gpg==1.16.0
GraKeL==0.1.10
gravbody==0.1.0
grpcio==1.70.0
grpcio-status==1.70.0
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib_metadata==8.2.0
iniconfig==1.1.1
iniparse==0.4
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
kiwisolver==1.4.4
lark==1.1.1
lark-parser==0.12.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.43.0
locket==1.0.0
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==3.0.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
mccabe==0.6.1
mercurial==6.1.1
mistune==3.0.2
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.46.0
natsort==8.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nemo-emblems==5.6.0
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
nose2==0.9.2
notebook_shim==0.2.4
notify2==0.3
ntlm-auth==1.5.0
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
onboard==1.4.1
overrides==7.7.0
packaging==24.1
PAM==0.4.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patiencediff==0.2.1
pbr==6.0.0
pexpect==4.8.0
Pillow==9.0.1
pipdeptree==2.23.1
platformdirs==4.2.2
plotly==5.23.0
pluggy==0.13.0
ply==3.11
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.11
pybind11==2.11.1
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.22
pycups==2.0.1
pycurl==7.44.1
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydeck==0.9.1
pydocstyle==6.1.1
pydot==1.4.2
pyelftools==0.27
pyflakes==2.4.0
pygame==2.6.0
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.10.1
pylatexenc==2.10
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gitlab==2.10.1
python-gnupg==0.4.8
python-json-logger==2.0.7
python-magic==0.4.24
python-xlib==0.29
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.1.0
qiskit==1.1.2
qiskit-terra==0.46.2
quantum-blackbird==0.5.0
quantum-xir==0.2.2
referencing==0.35.1
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-file==1.5.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==3.3
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rospkg-modules==1.6.0
rpds-py==0.20.0
rsa==4.9
rustworkx==0.15.1
scikit-learn==1.5.1
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
setproctitle==1.2.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.3.1
ssh-import-id==5.11
stack-data==0.6.3
stevedore==5.2.0
StrawberryFields==0.23.0
streamlit==1.46.1
streamlit-authenticator==0.4.2
symengine==0.11.0
sympy==1.14.0
systemd-python==234
tenacity==9.0.0
termcolor==2.4.0
terminado==0.18.1
thewalrus==0.21.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.1.1
tldextract==3.1.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.7.1
tornado==6.4.1
traitlets==5.14.3
triton==3.3.1
ttystatus==0.38
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
Unidecode==1.3.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.15.0
vcstool==0.3.0
vcstools==0.1.42
virtualenv==20.26.3
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.13.3
wsproto==1.0.0
wxPython==4.0.7
xanadu-cloud-client==0.3.2
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.2.0
youtube-dl==2021.12.17
zipp==1.0.0
